Get image_data from tl_content image subpalette(s)
This way, the available options are always compatible with the Contao
core. The motivation for this change is the fact that the field
`title` has been renamed to `imageTitle` in Contao 4.4, making the
field disappear from ElementGenerator content elements with image
fields. Changes like that will be reflected automatically with this
change.

Currently, creating a subpalette within the same ElementGenerator
field does not seem to be supported, which is why subpalettes of the
`useImage` subpalette are merged into the `useImage` subpalette. This
affects `overwriteMeta` in Contao 4.4.